Move files from PR folder to the Forge folder
There are several thoughts to be had here.

## getPr.svelte.ts

This is a file that helps with accessing the forge services, giving a simpler interface.

- Should this be a more generic catch all `forgePreview.svelte.ts` as it is bridging the gap between services and svelte?
- Should this live in the base of `forge/` or should there be a `forge/pr/` folder which groups PR related things?

## prContents.svelte.ts

This file is tied strongly to the PrDetailsModal.svelte.

- Should this live in a folder with that `PrDetailsModal.svelte` file
- Should this be more library like?

## Conclusion...

![](https://media1.giphy.com/media/v1.Y2lkPTc5MGI3NjExeDJrNWFnd2k5djJncmN3aWx2Zjdqemdta2ZsNTdvbnNlZ2RzM3ZqbCZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/K6VhXtbgCXqQU/giphy.gif)

There are lots of permutations, and I think for now, I'm content to just put them in the forge folder and visit these questions later as there are pros and cons with all of the options
