+++ apps/desktop/src/components/PrTemplateSection.svelte
@@ -2,7 +2,7 @@
 	import Select from '$components/Select.svelte';
 	import SelectItem from '$components/SelectItem.svelte';
 	import { getForge } from '$lib/forge/interface/forge';
-	import { TemplateService } from '$lib/pr/templateService';
+	import { TemplateService } from '$lib/forge/templateService';
 	import { Project } from '$lib/project/project';
 	import { getContext } from '@gitbutler/shared/context';
 	import { persisted } from '@gitbutler/shared/persisted';
