+++ apps/desktop/src/routes/[projectId]/+layout.ts
@@ -9,12 +9,12 @@ import { CommitDragActionsFactory } from '$lib/commits/dragActions.js';
 import { StackingReorderDropzoneManagerFactory } from '$lib/dragging/stackingReorderDropzoneManager';
 import { FetchSignal } from '$lib/fetchSignal/fetchSignal.js';
 import { UncommitedFilesWatcher } from '$lib/files/watcher';
+import { TemplateService } from '$lib/forge/templateService';
 import { HistoryService } from '$lib/history/history';
 import { StackPublishingService } from '$lib/history/stackPublishingService';
 import { SyncedSnapshotService } from '$lib/history/syncedSnapshotService';
 import { ProjectMetrics } from '$lib/metrics/projectMetrics';
 import { ModeService } from '$lib/mode/modeService';
-import { TemplateService } from '$lib/pr/templateService';
 import { type Project } from '$lib/project/project';
 import { ProjectService } from '$lib/project/projectService';
 import { UpstreamIntegrationService } from '$lib/upstream/upstreamIntegrationService';
